organizing portfolio models and adding necessary fields
