+++ src/app/Financial/Wallet/CashReceipts/Model/Client.ts
@@ -1,0 +1,13 @@
+/**
+ * Representa la entidad Cliente con su información esencial.
+ * Utilizado para la selección de clientes y la obtención de datos básicos.
+ */
+export interface Client {
+    id: number;
+    name: string;
+    accountsReceivableAccount: {
+        id: number;
+        code: string; 
+        name: string;
+    };
+}
