+++ src/app/Financial/Wallet/CashReceipts/Model/DropdownOption.ts
@@ -1,0 +1,8 @@
+/**
+ * Modelo genérico para poblar listas desplegables (dropdowns) en la aplicación.
+ * Facilita la consistencia en componentes como p-dropdown.
+ */
+export interface DropdownOption {
+    label: string;
+    value: any; 
+}
