+++ src/app/Financial/Wallet/CashReceipts/Model/ReceiptDetail.ts
@@ -1,3 +1,7 @@
+/**
+ * Representa el detalle de un pago aplicado a una factura específica.
+ * Es un componente fundamental del objeto `ReceiptCreateRequest`.
+ */
 export interface ReceiptDetail {
     invoiceId: number;
     amountPaid: number;
