--- src/app/Financial/Wallet/CashReceipts/Model/VoidReceiptRequest.ts
@@ -1,3 +1,0 @@
-export interface VoidReceiptRequest {
-    reason: string;
-}
