+++ src/app/Financial/Wallet/CashReceipts/Model/api/VoidReceiptRequest.ts
@@ -1,0 +1,7 @@
+/**
+ * Define el payload necesario para el endpoint de la API
+ * que anula un recibo de caja.
+ */
+export interface VoidReceiptRequest {
+    reason: string;
+}
