+++ src/app/Financial/Wallet/CashReceipts/Model/api/index.ts
@@ -1,0 +1,3 @@
+export * from './ReceiptCreateRequest';
+export * from './ReceiptResponse';
+export * from './VoidReceiptRequest';
