+++ src/app/Financial/Wallet/CashReceipts/Model/index.ts
@@ -1,0 +1,6 @@
+export * from './Client';
+export * from './AccountinEntryLine';
+export * from './DropdownOption';
+export * from './Invoice';
+export * from './Receipt';
+export * from './ReceiptDetail';
