+++ src/app/Financial/Wallet/CashReceipts/Model/view/index.ts
@@ -1,0 +1,2 @@
+export * from './ReceiptListItem';
+export * from './ReceiptView';
